requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
twilio==7.16.2
urllib3==1.26.14
